+++ hutool-core/src/main/java/cn/hutool/core/io/unit/DataSize.java
@@ -28,7 +28,7 @@ public final class DataSize implements Comparable<DataSize> {
 	/**
 	 * The pattern for parsing.
 	 */
-	private static final Pattern PATTERN = Pattern.compile("^([+-]?\\d+(\\.\\d+)?)([a-zA-Z]{0,2})$");
+	private static final Pattern PATTERN = Pattern.compile("^([+-]?\\d+(\\.\\d+)?)([a-zA-Z]{0,3})$");
 
 	/**
 	 * Bytes per Kilobyte(KB).